Initial commit.
The development of RT3DReconst_GUI started out in Visual Studio
to enable the use of NSight tools for CUDA. Now integrating with Qt.

The GUI receives an ExtdFrame from the ICEbot_GUI. The image is
transformed in space according to the EM reading. The non-masked points
in the image are added to a grid. Gridding is performed on the GPU.

Need to implement the interpolation kernel for gap filling between frames.

Signed-off-by: Degirmenci, Alperen <adegirmenci@seas.harvard.edu>